Split downloader_utils.py into separate files for WDI and PWT data sources
